# MAINT: speed up rvs of nakagami in scipy.stats