#### Reference issue
gh-10106

#### What does this implement/fix?
- speed up `stats.nakagami.rvs` by using that rvs can be written as sqrt of Gamma rvs (see https://en.wikipedia.org/wiki/Nakagami_distribution#Generation)
- add reference to Wikipedia and original article

#### Additional information

The speed-up is of order ~20x.

The following snippet of code can be run in a notebook as a check:
```
import numpy as np
from scipy import stats
from scipy import special as sc
import matplotlib.pyplot as plt

rng = np.random.default_rng()

nu = 2.3
fig, ax = plt.subplots(1, 1)
x = np.linspace(stats.nakagami.ppf(0.01, nu),
                stats.nakagami.ppf(0.99, nu), 100)

ax.plot(x, stats.nakagami.pdf(x, nu), 'r-', lw=5, alpha=0.6)
r = np.sqrt(rng.standard_gamma(nu, size=10_000) / nu)
ax.hist(r, density=True, bins=20, histtype='stepfilled', alpha=0.2)
plt.show()

N = 100_000
q = rng.uniform(size=N)
for nu in [0.5, 1.5, 3.5, 10]:
    print(nu)
    %timeit np.sqrt(rng.standard_gamma(nu, size=N) / nu)
    %timeit np.sqrt(1.0/nu*sc.gammaincinv(nu, q)) # this is the implementation of rvs in main
```

Output of timeit:
```
0.5
25.5 ms ± 1.85 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
579 ms ± 169 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
1.5
14.7 ms ± 1.63 ms per loop (mean ± std. dev. of 7 runs, 100 loops each)
453 ms ± 98 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
3.5
16.8 ms ± 1.68 ms per loop (mean ± std. dev. of 7 runs, 100 loops each)
448 ms ± 96 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
10
13.7 ms ± 2.84 ms per loop (mean ± std. dev. of 7 runs, 100 loops each)
308 ms ± 47.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```