@tirthasheshpatel 2022-01-05
> the failures in the CI are related to this change.

Yes, I just noticed. Sorry for not noticing before! I will review this in more detail.