@chrisb83 2022-01-05
Hm, it looks like gh-13396 did not fix the issues with the fit method of nakagami in a robust way. Given that the rvs changed in this PR, `test_fit` of `TestNakagami` in test_distributions.py fails. 

```
@pytest.mark.parametrize('nu', [1.6, 2.5, 3.9])
    @pytest.mark.parametrize('loc', [25.0, 10, 35])
    @pytest.mark.parametrize('scale', [13, 5, 20])
    def test_fit(self, nu, loc, scale):
        # Regression test for gh-13396 (21/27 cases failed previously)
        # The first tuple of the parameters' values is discussed in gh-10908
        N = 100
        samples = stats.nakagami.rvs(size=N, nu=nu, loc=loc,
                                                        scale=scale, random_state=1337)
        nu_est, loc_est, scale_est = stats.nakagami.fit(samples)
        assert_allclose(nu_est, nu, rtol=0.2)
        assert_allclose(loc_est, loc, rtol=0.2)
        assert_allclose(scale_est, scale, rtol=0.
```  

But if I change the seed, the test frequently fails on master, too. Small experiment below:

```
import numpy as np
from scipy import stats
import itertools


for version in ['new', 'master']:
    fails, counter = 0, 0
    for seed in [345, 53648, 46437, 25426, 43673, 122]:
        for args in itertools.product((1.6, 2.5, 3.9),
                                                     (25.0, 10, 35),
                                                     (13, 5, 20)):
            nu, loc, scale = args
            N = 100
            if version == 'master':
                samples = stats.nakagami.rvs(size=N, nu=nu, loc=loc,
                                                                scale=scale, random_state=seed)
            elif version == 'new':
                rng = np.random.default_rng(seed)
                samples = loc + scale * np.sqrt(rng.standard_gamma(nu, size=N) / nu)
            nu_est, loc_est, scale_est = stats.nakagami.fit(samples)
            fails += np.abs(nu_est - nu) / nu < 0.2
            fails += np.abs(loc_est - loc) / loc < 0.2
            fails += np.abs(scale_est - scale) / scale < 0.2
            counter += 3
    print(f'{version}: {fails} out of {counter} failed...')
    # new: 224 out of 486 failed...
    # master: 179 out of 486 failed...
```
so no matter which method is used, the fit is often not accurate...

So it might be that we have to reopen the issue gh-10908 and mark the test as xfail.