@tirthasheshpatel 2022-01-05
The failing tests (`scipy/stats/tests/test_distributions.py::TestNakagami::test_fit`) are _very_ sensitive to the seed and sample size. When I change the sample size from 100 -> 1000, the tests start failing on the main branch. These tests fail even with 100,000 samples [on the main branch]. Also, changing only the seed makes all 27/27 tests fail.

Moreover, I checked histograms of the samples generated by the RVS method implemented in this PR and it matches the PDF almost exactly for all the combinations of parameters in the failing tests. So, I don't think there is any issue with the `_rvs` method in this PR.

```python
In [1]: import numpy as np

In [2]: import matplotlib.pyplot as plt

In [3]: from scipy.stats import nakagami

In [4]: from itertools import product

In [5]: nus = [1.6, 2.5, 3.9]

In [6]: locs = [25, 10, 53]

In [7]: scales = [13, 5, 20]

In [8]: for nu, loc, scale in product(nus, locs, scales):
   ...:     rvs = nakagami.rvs(nu=nu, loc=loc, scale=scale, size=100)
   ...:     x = np.linspace(rvs.min()-0.1, rvs.max()+0.1, num=10000)
   ...:     fx = nakagami.pdf(x, nu=nu, loc=loc, scale=scale)
   ...:     plt.plot(x, fx)
   ...:     plt.hist(rvs, bins=10, density=True)
   ...:     plt.show()

```

`TestNakagami::test_fit` was added in #13396.

The tests pass locally both on main and with the changes in this PR after changing sample size from 100 to 10000 and loosening the absolute tolerance of the test checking if the derivatives wrt the parameters equal 0 to 1e-1. I think this tolerance and sample size make more sense. @mdhaber Since you were involved in reviewing #13396 and proposing the estimates for `_fitstart`, do you think it'd be OK to change the tolerances and sample size this way?