@tirthasheshpatel 2022-01-05
> So it might be that we have to reopen the issue [gh-10908](https://github.com/scipy/scipy/issues/10908) and mark the test as xfail

Marking them as xfail and reopening the issue also seems like a reasonable solution.