@tupui 2022-01-05
Should we then change the tests to have something more robust no? Based on the histogram for instance?