@chrisb83 2022-01-05
I used the wrong inequality when counting the fails: it should be 

```
fails += np.abs(nu_est - nu) / nu > 0.2
fails += np.abs(loc_est - loc) / loc > 0.2
fails += np.abs(scale_est - scale) / scale > 0.2
```

so i actually computed the cases that pass and the result reads:

new: 224 out of 486 passed...
master: 179 out of 486 passed...