@mdhaber 2022-01-06
We probably shouldn't have closed gh-10908. There was always the hope to add analytical fits for Nakagami. All we did here was give a [theoretically justified](https://github.com/scipy/scipy/pull/13396#discussion_r559247425) estimate of the guess for optimization.

I would xfail the tests rather than loosening the tolerances or changing the test method. `test_fit` checks the first-order conditions for optimality. In normal circumstances (e.g. unless we're at a constraint), they need to be satisfied or we haven't done MLE.