@tirthasheshpatel 2022-01-06
> If needed, I would xfail `TestNakagami::test_fit` rather than loosening the tolerances.

Sounds good. So, with changes in this PR, it would be necessary to mark these tests xfail. Is it OK to change that here or should we do it in a new PR?