@chrisb83 2022-01-06
> Sounds good. So, with changes in this PR, it would be necessary to mark these tests xfail. Is it OK to change that here or should we do it in a new PR?

I just reopened the issue and will mark the test xfail in this PR after work.