@chrisb83 2022-01-06
OK, CI is green after adding a line with xfail (except for a MacOS instance that did not complete the tests)