@chrisb83 2022-05-29
The notes contain a comment "faster random variate generation for ``gennorm`` and ``nakagami``". I think that is sufficient. what do you think?